Dev/skruk/script tests (#34)
* refactoring almost all markdown docs into docs folder
* added bash scripts test suite
* testing doc build
* updated docs to be alined with building, formatting, and linting
* testing agent sql content
* package script tests

+++ build_docs.sh
@@ -31,8 +31,16 @@ VERSION=$(grep 'VERSION =' build/_version.py | awk -F'"' '{print $2}')
 BUILD=$(grep 'BUILD =' build/_version.py | sed -E -e 's/^.*[=] ([0-9]+)\s*/\1/')
 CURRENT_DATE=$(date +%Y-%m-%d)
 
+rm _readme_full.md 2>/dev/null
+rm build/bom.yml 2>/dev/null
+
 PYTHONPATH="$PYTHONPATH:./src" python -m build.compile_bom
 PYTHONPATH="$PYTHONPATH:./src" python -m build.update_docs
+
+if [ ! -f _readme_full.md ] || [ ! -f build/bom.yml ]; then
+    echo "Error: Expected documentation files were not found" >&2
+    exit 10
+fi
 
 sed -E "s/# Dynatrace Snowflake Observability Agent$/# Dynatrace Snowflake Observability Agent (v$VERSION)\n<a id='dynatrace-snowflake-observability-agent'><\/a>/" _readme_full.md > _readme_full.tmp.md
 echo "" >>_readme_full.tmp.md
@@ -47,5 +55,10 @@ pandoc _readme_full.tmp.md \
     --metadata title=" "
 
 rm _readme_full.*
+
+npx prettier --write README.md
+npx prettier --write docs/PLUGINS.md
+npx prettier --write docs/SEMANTICS.md
+npx prettier --write docs/APPENDIX.md
 
 echo "Dynatrace-Snowflake-Observability-Agent-$VERSION.pdf file successfully created"
